refactor: reorganize zod schema files
