+++ src/validations/question/shared.ts
@@ -1,0 +1,6 @@
+import { createRequire } from "module";
+
+const require = createRequire(import.meta.url);
+const leoProfanity = require("leo-profanity");
+
+export { leoProfanity };
